Move preprocessing files into dedicated subdirectory
