+++ src/python/spectrseqtools/preprocessing/preprocessing.py
@@ -7,8 +7,8 @@ import yaml
 from clr_loader import get_mono
 
 from spectrseqtools.common import set_output_path
-from spectrseqtools.deconvolution import DeconvolutionParameters, DeisotopedPeakList
-from spectrseqtools.singleton_identification import RawPeakList
+from spectrseqtools.preprocessing.deconvolution import DeconvolutionParameters, DeisotopedPeakList
+from spectrseqtools.preprocessing.singleton_identification import RawPeakList
 
 rt = get_mono()
