+++ tests/test_deconvolution.py
@@ -1,7 +1,7 @@
 import pytest
 from clr_loader import get_mono
 
-from spectrseqtools.deconvolution import set_averagine
+from spectrseqtools.preprocessing.deconvolution import set_averagine
 
 rt = get_mono()
